requirements.txt
# requirements.txt
python-dotenv
solana
solders
httpx[http2]
pandas
joblib
aiosqlite
colorama